+++ app/src/main/java/com/example/weatherapp/data/models/MainWeatherModel.kt
@@ -3,10 +3,10 @@ package com.example.weatherapp.data.models
 import com.google.gson.annotations.SerializedName
 
 data class MainWeatherModel(
-    @SerializedName("temp") val temp: Float,
-    @SerializedName("feels_like") val feelsLike: Float,
-    @SerializedName("temp_min") val tempMin: Float,
-    @SerializedName("temp_max") val tempMax: Float,
+    @SerializedName("temp") val temp: Double,
+    @SerializedName("feels_like") val feelsLike: Double,
+    @SerializedName("temp_min") val tempMin: Double,
+    @SerializedName("temp_max") val tempMax: Double,
     @SerializedName("pressure") val pressure: Int,
     @SerializedName("humidity") val humidity: Int,
     @SerializedName("sea_level") val seaLevel: Int,
