+++ app/src/main/java/com/example/weatherapp/data/models/WindModel.kt
@@ -3,7 +3,7 @@ package com.example.weatherapp.data.models
 import com.google.gson.annotations.SerializedName
 
 data class WindModel(
-    @SerializedName("speed") val speed: Float,
+    @SerializedName("speed") val speed: Double,
     @SerializedName("deg") val deg: Int,
-    @SerializedName("gust") val gust: Float,
+    @SerializedName("gust") val gust: Double,
 )
