+++ app/src/main/java/com/example/weatherapp/domain/models/WeatherDetailsModel.kt
@@ -2,10 +2,10 @@ package com.example.weatherapp.domain.models
 
 data class WeatherDetailsModel(
     val name: String,
-    val temperatureInCelsius: Double,
-    val feelsLikeInCelsius: Double,
-    val maxTempInCelsius: Double,
-    val minTempInCelsius: Double,
+    val temperatureInCelsius: Int,
+    val feelsLikeInCelsius: Int,
+    val maxTempInCelsius: Int,
+    val minTempInCelsius: Int,
     val pressure: String,
     val humidity: String,
     val seaLevel: String,
