+++ app/src/main/java/com/example/weatherapp/domain/models/WindDetailsModel.kt
@@ -3,5 +3,5 @@ package com.example.weatherapp.domain.models
 data class WindDetailsModel(
     val speed: Double,
     val gust: Double,
-    val degrees: Double,
+    val degrees: Int,
 )
